harness-image-status: address PR 2 review feedback (http client reuse, local.go organization)
